requirements.txt
    python-dotenv>=1.0.0
    google-generativeai>=0.3.0

    # LangChain 0.0.350 compatible versions
langchain==0.1.20
langchain-community==0.0.38
langchain-core==0.1.52
langsmith==0.1.58

    packaging==23.2

    pydantic>=1.10.2
    requests>=2.28.0
    cffi==1.17.1
    cryptography==44.0.2
    pycparser==2.22